+++ changelog.md
@@ -7,6 +7,7 @@
 
 - Add support for MyST Markdown syntax for `click:example` and `click:run` Sphinx directives.
 - Add support for all `code-block` options to `click:example` and `click:run`: `:linenos:`, `:lineno-start:`, `:emphasize-lines:`, `:force:`, `:caption:`, `:name:`, `:class:` and `:dedent:`.
+- Add new `:show-source:`/`:hide-source:` and `:show-results:`/`:hide-results:` flag options to `click:example` and `click:run`. Closes {issue}`719`.
 - Support non-string choices in colored help screens. Closes {issue}`1284`.
 - Replace `LOG_LEVELS` mapping with `LogLevel` enum.
 - Remove `DEFAULT_LEVEL_NAME` constants.